JavaScript Modül Sistemi uygulandı. Projeyi Modül sistemine geçirme ve Datadan ürün listeme yapıldı.
