+++ js/data.json
@@ -1,87 +1,87 @@
 [
-    {
-        "id": 1,
-        "name": "Analogue Resin Strap",
-        "price": {
-            "oldPrice": 165.00,
-            "newPrice": 108.00
-        },
-        "discount": 22,
-        "img": {
-            "singleImage": "img/products/product1/1.png",
-            "thumbs": [
-                "img/products/product1/1.png",
-                "img/products/product1/2.png",
-                "img/products/product1/3.png"
-            ]
-        }
+  {
+    "id": 1,
+    "name": "Analogue Resin Strap",
+    "price": {
+      "oldPrice": 165.0,
+      "newPrice": 108.0
     },
-    {
-        "id": 2,
-        "name": "Ridley High Waist",
-        "price": {
-            "oldPrice": 165.00,
-            "newPrice": 208.00
-        },
-        "discount": 33,
-        "img": {
-            "singleImage": "img/products/product2/1.png",
-            "thumbs": [
-                "img/products/product2/1.png",
-                "img/products/product2/2.png",
-                "img/products/product2/3.png"
-            ]
-        }
-    },
-    {
-        "id": 3,
-        "name": "Blush Beanie",
-        "price": {
-            "oldPrice": 165.00,
-            "newPrice": 38.00
-        },
-        "discount": 90,
-        "img": {
-            "singleImage": "img/products/product3/1.png",
-            "thumbs": [
-                "img/products/product3/1.png",
-                "img/products/product3/2.png",
-                "img/products/product3/3.png"
-            ]
-        }
-    },
-    {
-        "id": 4,
-        "name": "Mercury Tee",
-        "price": {
-            "oldPrice": 165.00,
-            "newPrice": 48.00
-        },
-        "discount": 40,
-        "img": {
-            "singleImage": "img/products/product4/1.png",
-            "thumbs": [
-                "img/products/product4/1.png",
-                "img/products/product4/2.png",
-                "img/products/product4/3.png"
-            ]
-        }
-    },
-    {
-        "id": 5,
-        "name": "La Bohème Rose Gold",
-        "price": {
-            "oldPrice": 165.00,
-            "newPrice": 48.00
-        },
-        "discount": 40,
-        "img": {
-            "singleImage": "img/products/product5/1.png",
-            "thumbs": [
-                "img/products/product5/1.png",
-                "img/products/product5/2.png",
-                "img/products/product5/3.png"
-            ]
-        }
+    "discount": 22,
+    "img": {
+      "singleImage": "img/products/product1/1.png",
+      "thumbs": [
+        "img/products/product1/1.png",
+        "img/products/product1/2.png",
+        "img/products/product1/3.png"
+      ]
     }
+  },
+  {
+    "id": 2,
+    "name": "Ridley High Waist",
+    "price": {
+      "oldPrice": 165.0,
+      "newPrice": 208.0
+    },
+    "discount": 33,
+    "img": {
+      "singleImage": "img/products/product2/1.png",
+      "thumbs": [
+        "img/products/product2/1.png",
+        "img/products/product2/2.png",
+        "img/products/product2/3.png"
+      ]
+    }
+  },
+  {
+    "id": 3,
+    "name": "Blush Beanie",
+    "price": {
+      "oldPrice": 165.0,
+      "newPrice": 38.0
+    },
+    "discount": 90,
+    "img": {
+      "singleImage": "img/products/product3/1.png",
+      "thumbs": [
+        "img/products/product3/1.png",
+        "img/products/product3/2.png",
+        "img/products/product3/3.png"
+      ]
+    }
+  },
+  {
+    "id": 4,
+    "name": "Mercury Tee",
+    "price": {
+      "oldPrice": 165.0,
+      "newPrice": 48.0
+    },
+    "discount": 40,
+    "img": {
+      "singleImage": "img/products/product4/1.png",
+      "thumbs": [
+        "img/products/product4/1.png",
+        "img/products/product4/2.png",
+        "img/products/product4/3.png"
+      ]
+    }
+  },
+  {
+    "id": 5,
+    "name": "La Bohème Rose Gold",
+    "price": {
+      "oldPrice": 165.0,
+      "newPrice": 48.0
+    },
+    "discount": 40,
+    "img": {
+      "singleImage": "img/products/product5/1.png",
+      "thumbs": [
+        "img/products/product5/1.png",
+        "img/products/product5/2.png",
+        "img/products/product5/3.png"
+      ]
+    }
+  }
 ]
