+++ js/main.js
@@ -1,0 +1,17 @@
+import headerFunc from "./header.js";
+import productsFunc from "./products.js";
+
+
+// ! Add Product to Locak Stroge
+
+async function getData() {
+    const photos = await fetch("../js/data.json");
+    const data = await photos.json();
+
+    data ? localStorage.setItem("products", JSON.stringify(data)) : [];
+}
+
+getData();
+
+const products = localStorage.getItem("products");
+console.log(JSON.parse(products));
